+++ rust-toolchain
@@ -1,0 +1,1 @@
+nightly
